+++ mic_test_reactjs/functions/APIs/users.js
@@ -1,8 +1,24 @@
-const { db } = require("../db");
+const db = require("../db");
 const express = require("express");
 const router = express.Router();
-const { check, validationResult, body } = require("express-validator");
 
+const config = {
+  apiKey: "AIzaSyDNV3VNJxiqxcGWJJv3-ClrNSplSMnkJ4A",
+  authDomain: "mictest-7dd37.firebaseapp.com",
+  databaseURL: "https://mictest-7dd37.firebaseio.com",
+  projectId: "mictest-7dd37",
+  storageBucket: "mictest-7dd37.appspot.com",
+  messagingSenderId: "898308860024",
+  appId: "1:898308860024:web:6a614630c965eecf2fa58d",
+  measurementId: "G-HDN86Q140W",
+};
+
+const firebase = require("firebase");
+firebase.initializeApp(config);
+
+
+
+const { check, validationResult, body } = require("express-validator");
 router.post(
   "/register",
   [
@@ -26,7 +42,6 @@ router.post(
       confirmPassword,
       handle,
     };
-
     const errors = validationResult(req);
     if (!errors.isEmpty()) {
       return res.status(400).json({ errors: errors.array() });
@@ -37,9 +52,7 @@ router.post(
       if (doc.exists) {
         return res.status(400).json({ handle: "this handle is already taken" });
       }
-      const data = await firebase
-        .auth()
-        .createUserWithEmailAndPassword(newUser.email, newUser.password);
+      const data = await firebase.auth().createUserWithEmailAndPassword(newUser.email, newUser.password);
       const userId = await data.user.uid;
       const token = await data.user.getIdToken();
       const userCredentials = {
@@ -52,6 +65,7 @@ router.post(
       await db.doc(`/users/${newUser.handle}`).set(userCredentials);
       return await res.status(201).json({ token });
     } catch (err) {
+      console.log("May",err)
       if (err.code === "auth/email-already-in-use") {
         return res.status(400).json({ email: "Email already in user" });
       }
