+++ mic_test_reactjs/functions/db.js
@@ -1,6 +1,6 @@
 const admin = require("firebase-admin"); // ETO
-admin.initializeApp();
 
+admin.initializeApp();
 const db = admin.firestore(); // ETO NEED KO
 
 module.exports = db;
